Added player form final version

+++ playerApp/src/app/player/pages/crud/crud.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { PlayerI } from 'src/app/shared/interfaces/player-i';
 import { ModalCreateComponent } from '../../modals/modal-create.component';
 import { PlayerService } from '../../services/player.service';
-
+import { faTrash, IconDefinition } from '@fortawesome/free-solid-svg-icons';
 @Component({
   selector: 'app-crud',
   templateUrl: './crud.component.html',
@@ -12,6 +12,7 @@ import { PlayerService } from '../../services/player.service';
 })
 export class CrudComponent implements OnInit {
   public players: PlayerI[] = [];
+  public trashIcon: IconDefinition = faTrash;
   public isLoading: boolean = true;
   constructor(private playerService: PlayerService, private router: Router, private modalCreate: NgbModal) { }
 
@@ -32,21 +33,34 @@ export class CrudComponent implements OnInit {
       })
   }
 
-  public onMoreDetails(playerId: string): void {
-    this.router.navigateByUrl(`/player/${playerId}`);
+  public onMoreDetails(playerId: string = ""): void {
+    if (playerId != "")
+      this.router.navigateByUrl(`/player/${playerId}`);
   }
 
-  public onDelete(playerId: string): void {
-    this.playerService.deleteById(playerId).subscribe(() => {
-      this.loadData();
-    },
-      (error: Error) => {
-        console.log(error);
-      })
+  public onDelete(playerId: string = ""): void {
+    if (playerId != "")
+      this.playerService.deleteById(playerId).subscribe(() => {
+        this.loadData();
+      },
+        (error: Error) => {
+          console.log(error);
+        })
   }
 
   public onAddPlayer(): void {
-    this.modalCreate.open(ModalCreateComponent, { size: 'xl' });
+    const modalCreate: NgbModalRef = this.modalCreate.open(ModalCreateComponent, { size: 'lg' });
+    modalCreate.result.then((playerData: PlayerI) => {
+      if (playerData) {
+        this.playerService.create(playerData).subscribe((newPlayer: PlayerI) => {
+          this.loadData();
+          // @TODO: Add Toast config
+        },
+          (error: Error) => {
+            console.log(error);
+          })
+      }
+    })
   }
 
 }
